+++ internal/resource/config/certificates/groups/certificates_group_resource_gen.go
@@ -85,7 +85,7 @@ func (r *certificatesGroupResource) Schema(ctx context.Context, req resource.Sch
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.RequiresReplace(),
 				},
-				Description: "Cryptographic Provider. This is only applicable if Hybrid HSM mode is true. Options are `LOCAL` or `HSM`.",
+				Description: "Cryptographic Provider. This is only applicable if Hybrid HSM mode is true. Options are `LOCAL` or `HSM`. This field is immutable and will trigger a replacement plan if changed.",
 				Validators: []validator.String{
 					stringvalidator.OneOf(
 						"LOCAL",
@@ -99,7 +99,7 @@ func (r *certificatesGroupResource) Schema(ctx context.Context, req resource.Sch
 			},
 			"file_data": schema.StringAttribute{
 				Required:    true,
-				Description: "The certificate data in PEM format. New line characters should be omitted or encoded in this value.",
+				Description: "The certificate data in PEM format. New line characters should be omitted or encoded in this value. This field is immutable and will trigger a replacement plan if changed.",
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.RequiresReplace(),
 				},
@@ -110,7 +110,7 @@ func (r *certificatesGroupResource) Schema(ctx context.Context, req resource.Sch
 			"group_id": schema.StringAttribute{
 				Optional:    true,
 				Computed:    true,
-				Description: "The persistent, unique ID for the certificate. It can be any combination of `[a-z0-9._-]`. This property is system-assigned if not specified.",
+				Description: "The persistent, unique ID for the certificate. It can be any combination of `[a-z0-9._-]`. This property is system-assigned if not specified. This field is immutable and will trigger a replacement plan if changed.",
 				PlanModifiers: []planmodifier.String{
 					stringplanmodifier.RequiresReplace(),
 					stringplanmodifier.UseStateForUnknown(),
@@ -128,7 +128,7 @@ func (r *certificatesGroupResource) Schema(ctx context.Context, req resource.Sch
 				Validators: []validator.String{
 					stringvalidator.LengthAtLeast(1),
 				},
-				Description: "Name of the group to manage certificates for.",
+				Description: "Name of the group to manage certificates for. This field is immutable and will trigger a replacement plan if changed.",
 			},
 			"issuer_dn": schema.StringAttribute{
 				Computed:    true,