+++ internal/resource/config/sp/adapters/sp_adapter_resource.go
@@ -53,5 +53,5 @@ func (r *spAdapterResource) ModifyPlan(ctx context.Context, req resource.ModifyP
 	plan.Configuration, respDiags = pluginconfiguration.MarkComputedAttrsUnknownOnChange(plan.Configuration, state.Configuration)
 	resp.Diagnostics.Append(respDiags...)
 
-	resp.Plan.Set(ctx, plan)
+	resp.Diagnostics.Append(resp.Plan.Set(ctx, plan)...)
 }